17.3
17.3

+++ functions-admin.php
@@ -1451,7 +1451,8 @@ function battleplan_admin_menu() {
 	
 	if ( _USER_LOGIN == "battleplanweb" ) :
 		add_submenu_page( 'index.php', 'Clear ALL', 'Clear ALL', 'manage_options', 'clear-all', 'battleplan_clear_all' );	
-		add_submenu_page( 'index.php', 'Clear HVAC', 'Clear HVAC', 'manage_options', 'clear-hvac', 'battleplan_clear_hvac' );	
+		add_submenu_page( 'index.php', 'Clear HVAC', 'Clear HVAC', 'manage_options', 'clear-hvac', 'battleplan_clear_hvac' );			
+		add_submenu_page( 'index.php', 'Launch Site', 'Launch Site', 'manage_options', 'launch-site', 'battleplan_launch_site' );	
 		add_submenu_page( 'index.php', 'Run Chron', 'Run Chron <div class="admin-note">'.$chronTime.'</div>', 'manage_options', 'run-chron', 'battleplan_force_run_chron' );	
 		add_submenu_page( 'index.php', 'Site Audit', 'Site Audit <div class="admin-note">'.date("F j, Y", (int)$siteUpdated).'</div>', 'manage_options', 'site-audit', 'battleplan_site_audit' );	
 	endif;
@@ -2127,62 +2128,69 @@ function battleplan_site_audit() {
 
 // Set up brand new site
 function battleplan_clear_all() {
-	$products = get_posts( array('post_type'=>'products', 'numberposts'=>-1) );
-	foreach ($products as $post) wp_delete_post( $post->ID, true );
-
-	$args = array( 'post_status' => 'inherit', 'posts_per_page' => -1, 'post_type' => 'attachment', 'post_mime_type' => 'image', );
-	$args['tax_query'] = array( array( 'taxonomy' => 'image-categories', 'terms' => 'products', 'field' => 'slug', ),);
-	$getImg = new WP_Query( $args );
-
-	if ( $getImg->have_posts() ) : 
-		while ( $getImg->have_posts() ) :
-			$getImg->the_post();		
-			wp_delete_attachment( get_the_ID(), true );
-		endwhile; 
-	endif;
-
-	wp_reset_postdata();	
-	battleplan_clear_hvac();
+	battleplan_clear_hvac(true);
 }
 
-function battleplan_clear_hvac() {
+function battleplan_clear_hvac($all=false) {
+	$deleteImgs = array ('testimonials', 'photos', 'graphics', 'logos');
+	$keepPages = array ('home', 'contact-us', 'product-overview');
+
 	$optimized = get_posts( array('post_type'=>'optimized', 'numberposts'=>-1) );
 	$testimonials = get_posts( array('post_type'=>'testimonials', 'numberposts'=>-1) );
 	$galleries = get_posts( array('post_type'=>'galleries', 'numberposts'=>-1) );
 	$posts = get_posts( array('post_type'=>'post', 'numberposts'=>-1) );
+	$pages = get_posts( array('post_type'=>'page', 'numberposts'=>-1) );
+	
+	if ( $all == true ) :
+		$products = get_posts( array('post_type'=>'products', 'numberposts'=>-1) );
+		foreach ($products as $post) wp_delete_post( $post->ID, true );
+		array_push($deleteImgs, 'products');
+		if (in_array('product-overview', $keepPages)) unset($keepPages[array_search('product-overview', $keepPages)]);
+	endif;
+
 	foreach ($optimized as $post) wp_delete_post( $post->ID, true );
 	foreach ($testimonials as $post) wp_delete_post( $post->ID, true );
 	foreach ($galleries as $post) wp_delete_post( $post->ID, true );
 	foreach ($posts as $post) wp_delete_post( $post->ID, true );
+	foreach ($pages as $post) if ( !in_array( $post->post_name, $keepPages) ) wp_delete_post( $post->ID, true );
 
 	$args = array( 'post_status' => 'inherit', 'posts_per_page' => -1, 'post_type' => 'attachment', 'post_mime_type' => 'image', );
-	$terms = array ('testimonials', 'photos', 'graphics', 'logos');
-	$args['tax_query'] = array( array( 'taxonomy' => 'image-categories', 'terms' => $terms, 'field' => 'slug', ),);
+	$args['tax_query'] = array( array( 'taxonomy' => 'image-categories', 'terms' => $deleteImgs, 'field' => 'slug', ),);
 	$getImg = new WP_Query( $args );
 
 	if ( $getImg->have_posts() ) : 
 		while ( $getImg->have_posts() ) :
 			$getImg->the_post();	
-			$keepImg = array( 'logo.png', 'logo.webp', 'site-icon.png', 'site-icon.webp', 'favicon.png', 'favicon.webp');
-			
+			$keepImg = array( 'logo.png', 'logo.webp', 'site-icon.png', 'site-icon.webp', 'favicon.png', 'favicon.webp');			
 			if ( !in_array( basename( get_attached_file( get_the_ID() )), $keepImg) ) wp_delete_attachment( get_the_ID(), true );
 		endwhile; 
 		wp_reset_postdata();
 	endif;
-
-	updateOption('bp_chron_time', 0);
-	updateOption('bp_launch_date', date('Y-m-d'));
 	
+	header("Location: /wp-admin/");
+	exit();
+}  
+
+function battleplan_launch_site() {
 	delete_option('bp_gbp_update');		
 	delete_option('bp_site_hits_ga4');	
 	delete_option('bp_site_hits_ua_1');	
 	delete_option('bp_site_hits_ua_2');		
 	delete_option('bp_site_hits_ua_3');		
 	delete_option('bp_site_hits_ua_4');		
-	delete_option('bp_site_hits_ua_5');	
+	delete_option('bp_site_hits_ua_5');
+	delete_option('bp_site_hits_ua_1_backup');	
+	delete_option('bp_site_hits_ua_2_backup');		
+	delete_option('bp_site_hits_ua_3_backup');		
+	delete_option('bp_site_hits_ua_4_backup');		
+	delete_option('bp_site_hits_ua_5_backup');	
 	delete_option('bp_site_audit_details');
+
+	updateOption('bp_chron_time', 0);
+	updateOption('bp_launch_date', date('Y-m-d'));
 	
 	header("Location: /wp-admin/");
 	exit();
 }  
+
 ?>
